# contracts/openapi/shopping-list.openapi.yaml
openapi: 3.0.3
info:
  title: shopping-list API
  description: >
    Формирование списка покупок по периоду со снимком ингредиентов (ADR 0002).
    Повторный build за тот же период — replace активного списка (политика MVP).
  version: 1.0.0
servers:
  - url: http://localhost:8085/api/shopping/v1
    description: Локальная разработка
...
paths:
...
  /lists/{listId}/lines:
    post:
      tags: [Lists]
      summary: Добавить ручную строку
      operationId: addManualLine
      security:
        - InternalAuth: []
        - TrustedUser: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualLineCreate'
      responses:
        '201':
          description: Строка добавлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShoppingLine'
        '404':
          description: Список не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
components:
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
  securitySchemes:
    InternalAuth:
      type: apiKey
      in: header
      name: X-Internal-Auth
    TrustedUser:
      type: apiKey
      in: header
      name: X-User-Id
  schemas:
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    ShoppingLine:
      type: object
      required: [lineId, displayName, purchased]
      properties:
        lineId:
          type: string
          format: uuid
        displayName:
          type: string
          description: Название позиции в снимке
        quantity:
          type: number
          nullable: true
        unit:
          type: string
          nullable: true
        productCategory:
          type: string
          nullable: true
        purchased:
          type: boolean
        mergeNote:
          type: string
          description: Если единицы несовместимы — текст для UI вместо суммы
        sourceRecipeIds:
          type: array
          items:
            type: string
            format: uuid
...
    ManualLineCreate:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        productCategory:
          type: string
...